requirements.txt
discord.py
discord.ext.commands
feedparser
dateutil